Add files via upload
Linear mixed models for my thesis on Shannon's index of alpha diversity and Hygienic behaviour scores. Each of the models was tested with just hive, and with replicate/collar nested in hive. None of the four models was significant, so I went with the nested design for both in order to maximize usage of the available data.